\name{sa_shp}
\Rdversion{1.0}
\alias{sa_shp}
\docType{data}
\title{
Example dataset to map Small Area Estimates
}
\description{
\code{SAEval_example_shp} contains a \code{sf} object to map the small area estimates.
}
\usage{
data(SAEval_example_shp)
}
\format{
\code{sa_shp} is a \code{sf} object with the shapefile for the \code{sa} domain.  
}

\examples{
# Load example data
data(SAEval_example_shp)

summary(sa_shp)

}
\keyword{datasets}
