\name{SAEval-package}
\alias{SAEval}
\title{The R SAEval Package}
\description{
\bold{SAEval} is an R package for diagnostic analysis of Small Area Estimation (SAE). It provide a set of tools for the evaluation of SAE with respect to the direct estimates.
}
\details{
Working with SAE it is good practice to compare different estimators to find the one with the best performance. This package contains functions for statistical calculation of diagnostic procedure aimed at evaluate the quality of the SAE. In detail, in the package are developed some methods originally proposed in Brown et al (2001) to check the quality of SAE.

Furthermore is possible to produce graphical tools that map the chosen indicator for a spatial analysis.

For a complete list of functions, use \code{library(help = "SAEval")}.
}

\author{
Developed by Andrea Fasulo
}